Move client update lock to ELC service
