+++ modules/service/src/lib.rs
@@ -1,3 +1,4 @@
+mod client_lock;
 mod elc;
 mod enclave;
 mod service;
